Harvest dahsboard polish pass.
