+++ package.json
@@ -7,6 +7,7 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
+    "@types/react-transition-group": "^4.4.0",
     "@walletconnect/web3-provider": "^1.3.1",
     "axios": "^0.20.0",
     "browserify": "^16.5.2",
