+++ src/components/farmPrice/FarmPrice.jsx
@@ -10,9 +10,9 @@ const BluePanel = styled.div`
   background-color: ${(props) => props.theme.style.blueBackground};
   color: ${(props) => props.theme.style.primaryFontColor};
   font-family: ${fonts.headerFont};
-  padding: 2.5rem .7rem 2.5rem .7rem;
+  padding: 2.5rem .7rem 3.5rem .7rem;
   margin-right: .5rem;
-  margin-top: 1.5rem;
+  margin-top: 1.8rem;
   border: ${(props) => props.theme.style.mainBorder};
   border-radius: .5rem;
   box-sizing: border-box;
